chore(nodejs): move schema into cdp/
Relocate src/schema/ (cyclotron, hogflow) to src/cdp/schema/ — all 38
importers live under src/cdp/, so it belongs in that domain. Pure
structural move: git renames plus import-path rewrites to the
~/cdp/schema/ alias (relative importers normalized to the alias too).
No logic changes.
