+++ nodejs/src/cdp/_tests/builders/hogflow.builder.ts
@@ -1,7 +1,7 @@
 import { randomUUID } from 'crypto'
 
+import { HogFlow, HogFlowAction, HogFlowEdge } from '~/cdp/schema/hogflow'
 import { findActionByType } from '~/cdp/services/hogflows/hogflow-utils'
-import { HogFlow, HogFlowAction, HogFlowEdge } from '~/schema/hogflow'
 import { logger } from '~/utils/logger'
 
 import { HOG_FILTERS_EXAMPLES } from '../examples'
