+++ nodejs/src/cdp/async-functions/customer_analytics.ts
@@ -1,6 +1,6 @@
 import { DateTime } from 'luxon'
 
-import { CyclotronInvocationQueueParametersFetchSchema } from '~/schema/cyclotron'
+import { CyclotronInvocationQueueParametersFetchSchema } from '~/cdp/schema/cyclotron'
 
 import { registerAsyncFunction } from '../async-function-registry'
