+++ nodejs/src/cdp/consumers/cdp-cyclotron-worker-hogflow-personhog.test.ts
@@ -2,8 +2,8 @@ import { createMockJobQueue } from '~/tests/helpers/mocks/job-queue.mock'
 
 import { DateTime } from 'luxon'
 
+import { HogFlow } from '~/cdp/schema/hogflow'
 import { PersonReadRepository } from '~/common/persons/repositories/person-repository'
-import { HogFlow } from '~/schema/hogflow'
 import { createCdpConsumerDeps } from '~/tests/helpers/cdp'
 import { getFirstTeam, resetTestDatabase } from '~/tests/helpers/sql'
