+++ nodejs/src/cdp/services/hogflows/actions/action.interface.ts
@@ -1,6 +1,7 @@
 import { DateTime } from 'luxon'
 
-import { HogFlowAction } from '../../../../schema/hogflow'
+import { HogFlowAction } from '~/cdp/schema/hogflow'
+
 import { CyclotronJobInvocationHogFlow, CyclotronJobInvocationResult } from '../../../types'
 import { HogExecutorExecuteAsyncOptions } from '../../hog-executor.service'
