+++ nodejs/src/cdp/services/hogflows/actions/delay.ts
@@ -1,6 +1,7 @@
 import { DateTime, DurationLike } from 'luxon'
 
-import { HogFlowAction } from '../../../../schema/hogflow'
+import { HogFlowAction } from '~/cdp/schema/hogflow'
+
 import { findContinueAction } from '../hogflow-utils'
 import { ActionHandler, ActionHandlerOptions, ActionHandlerResult } from './action.interface'
