+++ nodejs/src/cdp/services/hogflows/actions/random_cohort_branch.test.ts
@@ -1,7 +1,7 @@
 import { FixtureHogFlowBuilder } from '~/cdp/_tests/builders/hogflow.builder'
 import { createExampleHogFlowInvocation } from '~/cdp/_tests/fixtures-hogflows'
+import { HogFlowAction } from '~/cdp/schema/hogflow'
 import { CyclotronJobInvocationHogFlow } from '~/cdp/types'
-import { HogFlowAction } from '~/schema/hogflow'
 
 import { findActionById, findActionByType } from '../hogflow-utils'
 import { getRandomCohort } from './random_cohort_branch'
