+++ nodejs/src/cdp/services/hogflows/hogflow-duplicate-observer.service.ts
@@ -1,8 +1,8 @@
 import { Counter } from 'prom-client'
 
+import { HogFlowAction } from '~/cdp/schema/hogflow'
 import { RedisClient, RedisV2 } from '~/common/redis/redis-v2'
 
-import { HogFlowAction } from '../../../schema/hogflow'
 import { logger } from '../../../utils/logger'
 import { CyclotronJobInvocationHogFlow } from '../../types'
 import { mirrorCall } from '../../utils/mirror-call'
