+++ nodejs/src/cdp/services/hogflows/hogflow-executor.service.ts
@@ -1,7 +1,8 @@
 import { get } from 'lodash'
 import { DateTime } from 'luxon'
 
-import { HogFlow, HogFlowAction } from '../../../schema/hogflow'
+import { HogFlow, HogFlowAction } from '~/cdp/schema/hogflow'
+
 import { logger } from '../../../utils/logger'
 import { UUIDT } from '../../../utils/utils'
 import {
