+++ nodejs/src/cdp/services/hogflows/hogflow-quota-limiting.test.ts
@@ -1,5 +1,6 @@
+import { HogFlow } from '~/cdp/schema/hogflow'
+
 import { QuotaLimiting } from '../../../common/services/quota-limiting.service'
-import { HogFlow } from '../../../schema/hogflow'
 import { CyclotronJobInvocationHogFlow } from '../../types'
 import { HogFunctionMonitoringService } from '../monitoring/hog-function-monitoring.service'
 import {
