+++ nodejs/src/cdp/services/hogflows/hogflow-quota-limiting.ts
@@ -1,7 +1,8 @@
 import { Counter } from 'prom-client'
 
+import { HogFlow } from '~/cdp/schema/hogflow'
+
 import { QuotaLimiting } from '../../../common/services/quota-limiting.service'
-import { HogFlow } from '../../../schema/hogflow'
 import { CyclotronJobInvocationHogFlow } from '../../types'
 import { HogFunctionMonitoringService } from '../monitoring/hog-function-monitoring.service'
