+++ nodejs/src/cdp/services/messaging/recipient-preferences.service.test.ts
@@ -1,5 +1,6 @@
 import { FixtureHogFlowBuilder } from '~/cdp/_tests/builders/hogflow.builder'
 import { createExampleInvocation } from '~/cdp/_tests/fixtures'
+import { HogFlowAction } from '~/cdp/schema/hogflow'
 import { CyclotronJobInvocationHogFunction } from '~/cdp/types'
 import { getFirstTeam, resetTestDatabase } from '~/tests/helpers/sql'
 import { Hub, Team } from '~/types'
@@ -7,7 +8,6 @@ import { closeHub, createHub } from '~/utils/db/hub'
 import { logger } from '~/utils/logger'
 import { UUIDT } from '~/utils/utils'
 
-import { HogFlowAction } from '../../../schema/hogflow'
 import { RecipientsManagerService } from '../managers/recipients-manager.service'
 import { RecipientPreferencesService } from './recipient-preferences.service'
 import { RecipientTokensService } from './recipient-tokens.service'
